setsid SSH master process
Call `os.setsid()` on the SSH master process to prevent it from being
associated with the foreground process group. This ensures that when a
user presses Ctrl-C on the terminal, the SSH master process will not
receive any SIGINT so btrup can clean itself up properly through the SSH
connection.