backup: fix bug where it can't find a suitable common subvolume between src and dst
On the first run of backup(), where there are no snapshots on dst at
all, src will send the initial subvolume data over.

Here is the catch: this initial data **will not be stored as a snapshot
on dst**, but rather just as a regular readonly subvolume.

Our code, however, assumed that this initial data would be stored as a
snapshot subvolume. The result: on subsequent runs, we will never be
able to find a common subvolume on src and dst.

Fix this by not requiring dst_snapshots to be a snapshot.